External Response Model
